pyproject.toml
[project]
name = "asset-search-v2"
version = "0.1.0"
description = "Web-based physical asset discovery pipeline for corporate entities"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai-cloud-sdk",
    "firecrawl-py>=4.16.0",
    "litellm>=1.81.0",
    "instructor>=1.7.0",
    "pydantic>=2.12.0",
    "pydantic-ai>=1.58.0",
    "psycopg[binary]>=3.2.0",
    "duckdb>=1.2.0",
    "chonkie>=1.0.0",
    "cohere>=5.0.0",
    "tiktoken>=0.9.0",
    "rich>=14.0.0",
    "python-dotenv>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
